Add money-rails gem to handle price parsing
Since the msrp and unit_price are integers, I've made the assumption
that the integers represent cents. With that in mind, the money gem (and
its companion money-rails gem) turns integers into Money instances with
the standard 2 decimal places, plus provides a handy interface for
handling money calculations. This would be super useful for things like
calculating totals and such, but for now I'm just planning on using it
to show a nicely formatted price in the form and to ensure consistent
values between widget and line item.