Add order form using simple_form and cocoon
This adds the order form and the associated new/edit actions and routes.

I went with the simple_form and cocoon gems - simple_form because it's a
widely used form builder with a robust feature set, and cocoon because
it provides a simple jQuery implementation of a nested form with the ability to
add or remove as many sub-items as desired.

The line items' unit price is a bit tricky. Assuming we want it to be
based on the widget msrp/price, setting it in a before_validation might
be a reasonable solution. However, that makes all the validation specs
on line item blow up, because the instances shoulda_matchers makes don't
have widgets. To get around this, I used a before_save instead, but this
meant the line items throw validation errors, because before_save
callbacks happen after validations. Rails callbacks... there be dragons!

I think in reality a more robust solution would be better, perhaps using a
form object to handle creating the order + line items and setting the
unit price and/or more custom JS/jQuery, or even a JS framework, but
I think that's getting a bit beyond scope for a small code challenge.