Make order settings persist
I'm not quite clear what was meant by "can't be disabled once it's been
enabled", because with the previous implementation the `expedited?`
method would return a falsey value (nil) if it was set to false after setting
it to true. In any event, without persisting the settings to the
database, none of the settings will stick around anyway, so I added a
column to store them in the database.